move accelerator connector
