+++ pytorch_lightning/trainer/connectors/accelerator_connector.py
